requirements.txt
requests
beautifulsoup4
markdownify
pinecone
python-dotenv
fastapi
uvicorn
chainlit
openai
